Align torch.load and transformers imports for HF 4.46

+++ TTS/vc/modules/freevc/wavlm/wavlm.py
@@ -123,7 +123,7 @@ def compute_mask_indices(
             for length in sorted(lengths, reverse=True):
                 lens = np.fromiter(
                     (e - s if e - s >= length + min_space else 0 for s, e in parts),
-                    np.int,
+                    int,
                 )
                 l_sum = np.sum(lens)
                 if l_sum == 0:
